Moved redux actions and actionTypes
